+++ tests/unit/player/jump-markers/jump-marker-track.test.tsx
@@ -16,13 +16,16 @@ function makeController(durationS: number): PlayerController {
     durationS,
     isPlaying: false,
     isBuffering: false,
+    playbackRate: 1,
     activeSourceIndex: 0,
     getGameTimeS: () => 0,
     seekTo: () => {},
     seekBy: () => {},
+    stepBy: () => {},
     play: () => {},
     pause: () => {},
     togglePlay: () => {},
+    setPlaybackRate: () => {},
   };
 }